Fix buffer overflows in rgb_to_bgra
Both of these problems can be detected by running the tests with
RUSTFLAGS=-Zsanitizer=address.

When reading the last pixel in an image, we can't read any extra bytes.
Additionally, since the main loop reads past the end of the pixels it is
processing, it must never process the last pixel, even if width is an
exact multiple of ITEMS_PER_ITERATION.